Extracted some logic into separate classes
