+++ Service/ILocalisationFetcher.cs
@@ -1,0 +1,11 @@
+using System.Collections.Generic;
+
+using MoreCulturalNamesModBuilder.Service.Models;
+
+namespace MoreCulturalNamesModBuilder.Service
+{
+    public interface ILocalisationFetcher
+    {
+        IEnumerable<Localisation> GetGameLocationLocalisations(string locationGameId, string gameId);
+    }
+}
