+++ Service/INameNormaliser.cs
@@ -1,0 +1,7 @@
+namespace MoreCulturalNamesModBuilder.Service
+{
+    public interface INameNormaliser
+    {
+        string ToWindows1252(string name);
+    }
+}
